Merge to prod: 14 Oct 2021

+++ services/fileServices/MdPageServices/UnlinkedPageService.js
@@ -1,7 +1,11 @@
+const { BadRequestError } = require("@errors/BadRequestError")
+
 const {
   retrieveDataFromMarkdown,
   convertDataToMarkdown,
 } = require("@utils/markdown-utils")
+
+const { titleSpecialCharCheck } = require("@validators/validators")
 
 const UNLINKED_PAGES_DIRECTORY_NAME = "pages"
 
@@ -12,6 +16,8 @@ class UnlinkedPageService {
 
   async create(reqDetails, { fileName, content, frontMatter }) {
     // Ensure that third_nav_title is removed for files that are being moved from collections
+    if (titleSpecialCharCheck({ title: fileName, isFile: true }))
+      throw new BadRequestError("Special characters not allowed in file name")
     delete frontMatter.third_nav_title
     const newContent = convertDataToMarkdown(frontMatter, content)
     const { sha } = await this.gitHubService.create(reqDetails, {
@@ -62,6 +68,8 @@ class UnlinkedPageService {
     reqDetails,
     { oldFileName, newFileName, content, frontMatter, sha }
   ) {
+    if (titleSpecialCharCheck({ title: newFileName, isFile: true }))
+      throw new BadRequestError("Special characters not allowed in file name")
     const newContent = convertDataToMarkdown(frontMatter, content)
     await this.gitHubService.delete(reqDetails, {
       sha,
